test(gas-estimator): :art: refactor tests a tiny bit
also add TESTING.md docs and update bico SDK version
